+++ src/Tools.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spicyweb\tools;
+namespace spicyweb\oddsandends;
 
 use Craft;
 use craft\base\Plugin;
@@ -9,29 +9,27 @@ use craft\events\RegisterComponentTypesEvent;
 use craft\services\Dashboard;
 use craft\services\Fields;
 use craft\services\Plugins;
-use spicyweb\tools\fields\Ancestors as AncestorsField;
-use spicyweb\tools\fields\AuthorInstructions as AuthorInstructionsField;
-use spicyweb\tools\fields\CategoriesMultipleGroups as CategoriesMultipleGroupsField;
-use spicyweb\tools\fields\CategoriesSearch as CategoriesSearchField;
-use spicyweb\tools\fields\DisabledCategories as DisabledCategoriesField;
-use spicyweb\tools\fields\DisabledDropdown as DisabledDropdownField;
-
-use spicyweb\tools\fields\DisabledEntries as DisabledEntriesField;
-
-use spicyweb\tools\fields\DisabledLightswitch as DisabledLightswitchField;
-use spicyweb\tools\fields\DisabledNumber as DisabledNumberField;
-use spicyweb\tools\fields\DisabledPlainText as DisabledPlainTextField;
-use spicyweb\tools\fields\EntriesSearch as EntriesSearchField;
-use spicyweb\tools\fields\Grid as GridField;
-use spicyweb\tools\fields\Width as WidthField;
-use spicyweb\tools\widgets\RollYourOwn as RollYourOwnWidget;
-
+use spicyweb\oddsandends\fields\Ancestors as AncestorsField;
+use spicyweb\oddsandends\fields\AuthorInstructions as AuthorInstructionsField;
+use spicyweb\oddsandends\fields\CategoriesMultipleGroups as CategoriesMultipleGroupsField;
+use spicyweb\oddsandends\fields\CategoriesSearch as CategoriesSearchField;
+use spicyweb\oddsandends\fields\DisabledCategories as DisabledCategoriesField;
+use spicyweb\oddsandends\fields\DisabledDropdown as DisabledDropdownField;
+use spicyweb\oddsandends\fields\DisabledEntries as DisabledEntriesField;
+use spicyweb\oddsandends\fields\DisabledLightswitch as DisabledLightswitchField;
+use spicyweb\oddsandends\fields\DisabledNumber as DisabledNumberField;
+use spicyweb\oddsandends\fields\DisabledPlainText as DisabledPlainTextField;
+use spicyweb\oddsandends\fields\EntriesSearch as EntriesSearchField;
+use spicyweb\oddsandends\fields\Grid as GridField;
+use spicyweb\oddsandends\fields\Width as WidthField;
+use spicyweb\oddsandends\models\Settings;
+use spicyweb\oddsandends\widgets\RollYourOwn as RollYourOwnWidget;
 use yii\base\Event;
 
 /**
  * Class Tools
  *
- * @package spicyweb\tools
+ * @package spicyweb\oddsandends
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool
  * @since 2.0.0
@@ -89,6 +87,6 @@ class Tools extends Plugin
      */
     protected function createSettingsModel()
     {
-        return new \spicyweb\tools\models\Settings();
+        return new Settings();
     }
 }