+++ src/controllers/ToolsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spicyweb\tools\controllers;
+namespace spicyweb\oddsandends\controllers;
 
 use Craft;
 use craft\elements\Asset;
@@ -15,7 +15,7 @@ use craft\web\Controller;
 /**
  * Class ToolsController
  *
- * @package spicyweb\tools\controllers
+ * @package spicyweb\oddsandends\controllers
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool
  * @since 2.0.0