+++ src/fields/AuthorInstructions.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace spicyweb\tools\fields;
+namespace spicyweb\oddsandends\fields;
 
 use Craft;
 
 use craft\base\ElementInterface;
 use craft\base\Field;
 use craft\helpers\Template;
-use spicyweb\tools\assetbundles\tools\ToolsAsset;
+use spicyweb\oddsandends\assetbundles\tools\ToolsAsset;
 use yii\db\Schema;
 
 /**
  * AuthorInstructions Field
  *
- * @package spicyweb\tools\fields
+ * @package spicyweb\oddsandends\fields
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool
  * @since 2.0.0