+++ src/fields/DisabledDropdown.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spicyweb\tools\fields;
+namespace spicyweb\oddsandends\fields;
 
 use Craft;
 use craft\base\ElementInterface;
@@ -9,7 +9,7 @@ use craft\fields\Dropdown;
 /**
  * Disabled Dropdown Field
  *
- * @package spicyweb\tools\fields
+ * @package spicyweb\oddsandends\fields
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool
  * @since 2.0.0