+++ src/fields/DisabledPlainText.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spicyweb\tools\fields;
+namespace spicyweb\oddsandends\fields;
 
 use Craft;
 use craft\base\ElementInterface;
@@ -10,7 +10,7 @@ use yii\db\Schema;
 /**
  * Disabled Plain Text Field
  *
- * @package spicyweb\tools\fields
+ * @package spicyweb\oddsandends\fields
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool
  * @since 2.0.0