+++ src/fields/EntriesSearch.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace spicyweb\tools\fields;
+namespace spicyweb\oddsandends\fields;
 
 use Craft;
 
 use craft\base\ElementInterface;
 use craft\fields\Entries;
 use craft\helpers\Template;
-use spicyweb\tools\assetbundles\tools\ToolsAsset;
+use spicyweb\oddsandends\assetbundles\tools\ToolsAsset;
 
 /**
  * Entries Search Field
  *
- * @package spicyweb\tools\fields
+ * @package spicyweb\oddsandends\fields
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool
  * @since 2.0.0