+++ src/fields/data/GridData.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace spicyweb\tools\fields\data;
+namespace spicyweb\oddsandends\fields\data;
 
 /**
  * Class WidthData
  *
- * @package spicyweb\tools\fields\data
+ * @package spicyweb\oddsandends\fields\data
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool Ltd <naveed@supercooldesign.co.uk>
  * @since 2.1.0