+++ src/migrations/m220921_060212_update_developer_in_db.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace spicyweb\tools\migrations;
+namespace spicyweb\oddsandends\migrations;
 
 use Craft;
 use craft\db\Migration;
 use craft\db\Table;
-use spicyweb\tools\Tools;
-use spicyweb\tools\widgets\RollYourOwn;
+use spicyweb\oddsandends\Tools;
+use spicyweb\oddsandends\widgets\RollYourOwn;
 
 /**
  * m220921_060212_update_developer_in_db migration.
@@ -20,12 +20,15 @@ class m220921_060212_update_developer_in_db extends Migration
     {
         // Update field types
         $projectConfig = Craft::$app->getProjectConfig();
-        $newClassPrefix = preg_replace('/Tools$/', '', Tools::class);
-        $oldClassPrefix = preg_replace('/^spicyweb/', 'supercool', $newClassPrefix);
+        $newClassPrefix = preg_replace('/Plugin$/', '', Tools::class);
+        $oldClassPrefix = preg_replace('/^spicyweb\\\\oddsandends/', 'supercool\\tools', $newClassPrefix);
 
         foreach ($projectConfig->get('fields') ?? [] as $uid => $field) {
             if (strpos($field['type'], $oldClassPrefix) === 0) {
-                $projectConfig->set("fields.$uid.type", preg_replace('/^supercool/', 'spicyweb', $field['type']));
+                $projectConfig->set(
+                    "fields.$uid.type",
+                    preg_replace('/^supercool\\\\tools/', 'spicyweb\\oddsandends', $field['type'])
+                );
             }
         }
 
@@ -33,7 +36,7 @@ class m220921_060212_update_developer_in_db extends Migration
         $this->update(
             Table::WIDGETS,
             ['type' => RollYourOwn::class],
-            ['type' => preg_replace('/^spicyweb/', 'supercool', RollYourOwn::class)]
+            ['type' => preg_replace('/^spicyweb\\\\oddsandends/', 'supercool\\tools', RollYourOwn::class)]
         );
     }